requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
SQLAlchemy==2.0.22
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.4.2
PyMySQL==1.1.0